refactor: modules/guide에 가이드 관련 모듈 묶기
