+++ packages/web/src/app/[lng]/globals.css
@@ -1,4 +1,4 @@
-@import './memos/css/guide.css';
+@import '../../modules/guide/index.css';
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
