--- packages/web/src/app/[lng]/memos/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './guide';
