+++ packages/web/src/modules/guide/index.ts
@@ -1,1 +1,3 @@
 export { default as useGuide } from './useGuide';
+
+export * from './constant';
