+++ packages/web/src/modules/guide/useGuide.ts
@@ -1,6 +1,6 @@
 import { useDidMount } from '@extension/shared/hooks';
 import { checkLocalStorageTrue } from '@extension/shared/modules/local-storage';
-import { driverObj } from '../utils';
+import { driverObj } from './constant';
 
 export default function useGuide() {
   useDidMount(() => {
