+++ packages/web/tsconfig.json
@@ -15,5 +15,5 @@
     "target": "ES2015"
   },
   "include": ["src"],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", ".next"]
 }
